requirements.txt
black
pylint
pytest
click
databricks
databricks-cli
ipython
databricks-sql-connector
databricks-connect==10.4.*
fastapi
uvicorn[standard]
pandasql
tabulate
psycopg2.extras
pandas
